python_requirements.txt
Flask==1.1.1
Flask-Admin==1.5.4
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-WTF==0.14.2
MarkupSafe==1.1.1
psycopg2-binary==2.8.3
SQLAlchemy==1.3.10
SQLAlchemy-Utils==0.34.2
pylint-flask==0.6
atomicwrites==1.4.0
PyJWT==1.7.1
Flask-Markdown==0.3
